MNES-385 refactor ApplicationTreeItem and EntityTreeItem to remove duplicated code, update version of HierarchicalStructures Submodel
