+++ src/app/[locale]/viewer/_components/submodel-elements/entity-component/EntityTreeItem.tsx
@@ -1,12 +1,10 @@
 import * as React from 'react';
-import { TreeItem, TreeItemContentProps, TreeItemProps, useTreeItemState } from '@mui/x-tree-view';
+import { useTreeItemState } from '@mui/x-tree-view';
 import clsx from 'clsx';
-import Typography from '@mui/material/Typography';
-import { Box, Button, IconButton, styled, useTheme } from '@mui/material';
+import { Box, Button, IconButton } from '@mui/material';
 import { Entity, ISubmodelElement, KeyTypes, RelationshipElement } from '@aas-core-works/aas-core3.0-typescript/types';
 import { AssetIcon } from 'components/custom-icons/AssetIcon';
 import { ArrowForward, ArticleOutlined, InfoOutlined, PinDropOutlined } from '@mui/icons-material';
-import { IconCircleWrapper } from 'components/basics/IconCircleWrapper';
 import { FormattedMessage } from 'react-intl';
 import { messages } from 'lib/i18n/localization';
 import { useRouter } from 'next/navigation';
@@ -15,40 +13,15 @@ import { EntityDetailsDialog } from './EntityDetailsDialog';
 import { RelationShipDetailsDialog } from './RelationShipDetailsDialog';
 import { GetKeyType } from 'lib/util/KeyTypeUtil';
 import { useApis } from 'components/azureAuthentication/ApiProvider';
+import { CustomTreeItemContentProps, CustomTreeItemProps, ExpandableTreeitem, StyledTreeItem } from '../TreeItem';
 
-interface CustomTreeItemProps extends TreeItemProps {
-    data?: ISubmodelElement;
-}
 
-interface CustomTreeItemContentProps extends TreeItemContentProps {
-    data?: ISubmodelElement;
-}
-
-const StyledTreeItem = styled(TreeItem)(({ theme }) => ({
-    '.MuiTreeItem-content': {
-        userSelect: 'none',
-        margin: 0,
-        borderBottom: '1px solid',
-        borderColor: theme.palette.divider,
-        '&.Mui-focused': {
-            backgroundColor: 'transparent',
-        },
-        '&.Mui-focused:hover': {
-            backgroundColor: theme.palette.action.hover,
-        },
-        '&.Mui-focused.Mui-selected': {
-            backgroundColor: theme.palette.action.selected,
-        },
-    },
-}));
 
 const CustomContent = React.forwardRef(function CustomContent(props: CustomTreeItemContentProps, ref) {
     const navigate = useRouter();
-    const theme = useTheme();
-    const { classes, className, label, itemId, icon: iconProp, expansionIcon, displayIcon, data } = props;
+    const { classes, className, label, itemId, icon: iconProp, data, ...other } = props;
     const { disabled, expanded, selected, focused, handleExpansion } = useTreeItemState(itemId);
     const { discoveryServiceClient } = useApis();
-    const toggleIcon = iconProp || expansionIcon || displayIcon;
     const isEntity = GetKeyType(data as ISubmodelElement) === KeyTypes.Entity;
     const dataIcon = isEntity ? (
         <AssetIcon fontSize="small" color="primary" />
@@ -103,36 +76,14 @@ const CustomContent = React.forwardRef(function CustomContent(props: CustomTreeI
                 ref={ref as React.Ref<HTMLDivElement>}
                 data-testid="bom-entity"
             >
-                <Box data-testid="expand-entity-icon" className={classes.iconContainer} sx={{ py: 1 }}>
-                    {toggleIcon}
-                </Box>
-                <Box
-                    sx={{
-                        [theme.breakpoints.down(480)]: {
-                            flex: 1,
-                            overflow: 'hidden',
-                            textOverflow: 'ellipsis',
-                            whiteSpace: 'nowrap',
-                        },
-                    }}
-                    display="flex"
-                    alignItems="center"
-                >
-                    <IconCircleWrapper>{dataIcon}</IconCircleWrapper>
-                    <Typography
-                        component="div"
-                        sx={{
-                            py: 2,
-                            pr: 1,
-                            overflow: 'hidden',
-                            textOverflow: 'ellipsis',
-                            whiteSpace: 'nowrap',
-                        }}
-                        className={classes.label}
-                    >
-                        {label}
-                    </Typography>
-                </Box>
+                <ExpandableTreeitem 
+                    icon={iconProp} 
+                    dataIcon={dataIcon} 
+                    classes={classes}
+                    itemId={itemId}
+                    label={label}
+                    {...other}
+                />
                 <Box sx={{ ml: 'auto', pl: 1, display: 'flex' }}>
                     {assetId && !showDataDirectly && (
                         <>
