+++ src/lib/enums/SubmodelSemanticId.enum.ts
@@ -4,7 +4,7 @@ export enum SubmodelSemanticId {
     CarbonFootprintIRDI = '0173-1#01-AHE712#001',
     CarbonFootprint = 'https://admin-shell.io/idta/CarbonFootprint/CarbonFootprint/0/9',
     TimeSeries = 'https://admin-shell.io/idta/TimeSeries/1/1',
-    HierarchicalStructures = 'https://admin-shell.io/idta/HierarchicalStructures/1/0/Submodel',
+    HierarchicalStructures = 'https://admin-shell.io/idta/HierarchicalStructures/1/1/Submodel',
     Nameplate = 'https://admin-shell.io/zvei/nameplate/1/0/Nameplate',
     BillOfApplications = 'https://xitaso.com/BillOfApplications',
 }
